requirements.txt
anyio==3.6.2
apimatic-core==0.1.2
apimatic-core-interfaces==0.1.0
apimatic-requests-client-adapter==0.1.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
astroid==2.12.12
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.26.3
botocore==1.29.3
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
Clarify==0.4.0
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
daphne==4.0.0
datetime-truncate==1.1.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-allauth==0.51.0
django-client-side-image-cropping==0.1.10
django-import-export==3.0.1
django-letsencrypt==4.1.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.1
django-polymorphic==3.1.0
django-ratelimit==3.0.1
django-storages==1.13.1
django-tinymce==3.5.0
django4-background-tasks==1.2.7
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
enum34==1.1.10
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fqdn==1.5.1
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.19.3
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.10.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.10
jsonpickle==1.5.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keras==2.10.0
lazy-object-proxy==1.8.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.3
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.9
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydivert==2.1.0
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
redis==4.3.4
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
scipy==1.9.3
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2
sqlparse==0.4.3
squareup==23.0.0.20221019
stack-data==0.6.2
stripe==4.2.0
tablib==3.2.1
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.6
uri-template==1.2.0
urllib3==1.26.12
validators==0.20.0
virtualenv==20.16.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.8.1
psycopg2==2.9.5
zope.interface==5.5.1
